update:bug修改

+++ GameServer/AscensionProtocol/DTO/CricketStatusDTO.cs
@@ -10,18 +10,18 @@ namespace AscensionProtocol
     public class CricketStatusDTO
     {
         public virtual int CricketID { set; get; }
-        public virtual int Atk { set; get; }
-        public virtual int Hp { set; get; }
-        public virtual int Defense { set; get; }
-        public virtual int Mp { set; get; }
-        public virtual int MpReply { set; get; }
-        public virtual int Crt { set; get; }
-        public virtual int CrtAtk { set; get; }
-        public virtual int CrtDef { set; get; }
-        public virtual int ReduceAtk { set; get; }
-        public virtual int ReduceDef { set; get; }
-        public virtual int Rebound { set; get; }
-        public virtual int Eva { set; get; }
+        public virtual float Atk { set; get; }
+        public virtual float Hp { set; get; }
+        public virtual float Defense { set; get; }
+        public virtual float Mp { set; get; }
+        public virtual float MpReply { set; get; }
+        public virtual float Crt { set; get; }
+        public virtual float CrtAtk { set; get; }
+        public virtual float CrtDef { set; get; }
+        public virtual float ReduceAtk { set; get; }
+        public virtual float ReduceDef { set; get; }
+        public virtual float Rebound { set; get; }
+        public virtual float Eva { set; get; }
         public virtual double Speed { set; get; }
     }
 }
